src/main/java/pro/sky/java/course2/course2_lesson3/servise/CalculatorServiceImpl.java
package pro.sky.java.course2.course2_lesson3.servise;

import org.springframework.stereotype.Service;

@Service

public class CalculatorServiceImpl implements CalculatorService {
    @Override
    public int sum(int num1, int num2) {
        return num1 + num2;
    }

    @Override
    public int subtract(int num1, int num2) {
        return num1 - num2;
    }

    @Override
    public int multiply(int num1, int num2) {
        return num1 * num2;
    }

    @Override
    public int divide(int num1, int num2) {
        if (num2 == 0) {
            throw new IllegalArgumentException("Для 'num2' необходимо значение не равное нулю");
        }
        return num1 / num2;
    }

}
